test: isolate GitHub action pinning checks
